# cmd/api/main.go
package main

import (
	"context"
	"encoding/json"
	"log"
	"net/http"

	_ "github.com/RafayKhattak/aegis-iam-backend/docs"
	"github.com/RafayKhattak/aegis-iam-backend/internal/config"
	"github.com/RafayKhattak/aegis-iam-backend/internal/handlers"
	appMiddleware "github.com/RafayKhattak/aegis-iam-backend/internal/middleware"
	"github.com/RafayKhattak/aegis-iam-backend/internal/repository"
	"github.com/RafayKhattak/aegis-iam-backend/internal/services"
	appJWT "github.com/RafayKhattak/aegis-iam-backend/pkg/jwt"
	"github.com/go-chi/chi/v5"
	chiMiddleware "github.com/go-chi/chi/v5/middleware"
	httpSwagger "github.com/swaggo/http-swagger/v2"
)

type healthResponse struct {
	Status  string `json:"status"`
	Service string `json:"service"`
	Version string `json:"version"`
}

// @title Aegis-IAM API
// @version 1.0
// @description Identity and Access Management Service for Khattak Industries.
// @host localhost:8080
// @BasePath /
// @securityDefinitions.apikey BearerAuth
// @in header
// @name Authorization
func main() {
	appConfig := config.LoadConfig()
	ctx := context.Background()

	store, err := repository.NewStore(ctx, appConfig.DBSource)
	if err != nil {
		log.Fatalf("failed to initialize database store: %v", err)
	}
	defer store.Close()

	tokenManager := appJWT.NewTokenManager(appConfig.JWTSecret)
	userService := services.NewUserService(store, tokenManager, appConfig.TokenDuration)
	userHandler := handlers.NewUserHandler(userService)

	router := chi.NewRouter()
	router.Use(chiMiddleware.Logger)
	router.Use(chiMiddleware.Recoverer)
	router.Post("/users/register", userHandler.Register)
	router.Post("/users/login", userHandler.Login)
	router.Get("/swagger/*", httpSwagger.Handler(httpSwagger.URL("http://localhost:8080/swagger/doc.json")))
	router.Group(func(r chi.Router) {
		r.Use(appMiddleware.AuthMiddleware(tokenManager))
		r.Get("/users/me", userHandler.GetMe)
	})

	router.Get("/health", func(w http.ResponseWriter, r *http.Request) {
		w.Header().Set("Content-Type", "application/json")
		w.WriteHeader(http.StatusOK)

		response := healthResponse{
			Status:  "ok",
			Service: "aegis-iam",
			Version: "1.0.0",
		}

		if err := json.NewEncoder(w).Encode(response); err != nil {
			http.Error(w, http.StatusText(http.StatusInternalServerError), http.StatusInternalServerError)
		}
	})

	address := ":" + appConfig.Port
	log.Printf("Starting Aegis-IAM server on port %s...", appConfig.Port)
	if err := http.ListenAndServe(address, router); err != nil {
		log.Fatalf("server failed to start: %v", err)
	}
}